+++ astropy/io/fits/tests/test_header.py
@@ -2264,6 +2264,19 @@ class TestHeaderFunctions(FitsTestCase):
             else:
                 c.verify('exception')
 
+    def test_keyword_from_index(self):
+        '''
+        Testing private function for positive and negative index mapping to true correct index
+        '''
+
+        hdr = fits.Header([('MOONFRAC', 0.604829), ('MOONDIST', 82.5171118)])
+
+        hdr.append()
+        hdr.append()
+
+        assert hdr._keyword_from_index(1) == ('MOONDIST', 0)
+        assert hdr._keyword_from_index(-1) == ('', 1)   
+
 
 class TestRecordValuedKeywordCards(FitsTestCase):
     """